pdf_chatbot_gemini/requirements.txt
streamlit
python-dotenv
PyMuPDF
google-generativeai